Exposing model_selection in a similar way to scikit-learn(#3329)
Resolving #3267. It seems that model_selection is already properly exposed through `cuml.preprocessing.model_selection` A small test suite for `train_test_split` is included in this PR to demonstrate that it works as desired.

Authors:
  - Patrick Liu <patrick.liu01@gmail.com>

Approvers:
  - William Hicks (@wphicks)
  - John Zedlewski (@JohnZed)

URL: https://github.com/rapidsai/cuml/pull/3329
