+++ python/cuml/__init__.py
@@ -50,6 +50,7 @@ from cuml.manifold.umap import UMAP
 from cuml.metrics.accuracy import accuracy_score
 from cuml.metrics.cluster.adjusted_rand_index import adjusted_rand_score
 from cuml.metrics.regression import r2_score
+from cuml.model_selection import train_test_split
 
 from cuml.naive_bayes.naive_bayes import MultinomialNB
 
@@ -58,7 +59,6 @@ from cuml.neighbors.kneighbors_classifier import KNeighborsClassifier
 from cuml.neighbors.kneighbors_regressor import KNeighborsRegressor
 
 from cuml.preprocessing.LabelEncoder import LabelEncoder
-from cuml.preprocessing.model_selection import train_test_split
 
 from cuml.random_projection.random_projection import GaussianRandomProjection
 from cuml.random_projection.random_projection import SparseRandomProjection
