+++ python/cuml/model_selection/__init__.py
@@ -1,0 +1,4 @@
+from cuml.model_selection._split import train_test_split
+
+
+__all__ = ['train_test_split']
